+++ examples/w3a/src/WorldBuilder.js
@@ -20,19 +20,20 @@ export class WorldBuilder extends ModelService {
         for (let x = 0; x < Constants.sizeX; x++) {
             for (let y = 0; y < Constants.sizeY; y++) {
                 landMatrix[x][y][0] = Constants.voxel.lava;
-                // landMatrix[x][y][1] = Constants.voxel.rock;
-                // landMatrix[x][y][2] = Constants.voxel.rock;
-                // if (Math.random() < 0.7) landMatrix[x][y][2] = Constants.voxel.rock;
-                // if (Math.random() < 0.5) landMatrix[x][y][3] = Constants.voxel.dirt;
-                // if (Math.random() < 0.1) landMatrix[x][y][4] = Constants.voxel.dirt;
+                landMatrix[x][y][1] = Constants.voxel.rock;
+                if (Math.random() < 0.7) landMatrix[x][y][2] = Constants.voxel.rock;
+                if (Math.random() < 0.5) landMatrix[x][y][3] = Constants.voxel.dirt;
+                if (Math.random() < 0.1) landMatrix[x][y][4] = Constants.voxel.dirt;
 
 
             }
         }
-        landMatrix[2][2][1] = Constants.voxel.rock;
-        landMatrix[2][2][2] = Constants.voxel.rock;
-        landMatrix[2][2][3] = Constants.voxel.rock;
-        landMatrix[3][2][3] = Constants.voxel.rock;
+        // landMatrix[2][2][1] = Constants.voxel.dirt;
+        // landMatrix[2][2][2] = Constants.voxel.dirt;
+        // landMatrix[2][2][3] = Constants.voxel.dirt;
+        // landMatrix[3][2][3] = Constants.voxel.rock;
+        // landMatrix[4][2][3] = Constants.voxel.rock;
+        // landMatrix[5][2][3] = Constants.voxel.rock;
 
         // landMatrix[2][4][2] = Constants.voxel.rock;
         // landMatrix[4][2][2] = Constants.voxel.rock;